Reverted PlatformAliasTab to its original UI
