+++ lib/screens/alias_tab/components/empty_list_alias_tab.dart
@@ -1,0 +1,20 @@
+import 'package:anonaddy/shared_components/constants/material_constants.dart';
+import 'package:flutter/material.dart';
+
+class EmptyListAliasTabWidget extends StatelessWidget {
+  const EmptyListAliasTabWidget({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    final isDark = Theme.of(context).brightness == Brightness.dark;
+    return Center(
+      child: Text(
+        'It doesn\'t look like you have any aliases yet!',
+        style: Theme.of(context)
+            .textTheme
+            .bodyText1!
+            .copyWith(color: isDark ? Colors.white : kPrimaryColor),
+      ),
+    );
+  }
+}
